+++ bin/cli.php
@@ -33,6 +33,7 @@ use App\Console\MigrateMailLogTokens;
 use App\Console\MigrateIpCreatedDayIndex;
 use App\Console\MigrateDropMailLogColumns;
 use App\Console\MigrateDropMailLogIndexes;
+use App\Console\MigrateMapUserConstraints;
 
 use App\Console\MigrateDb;
 use App\Console\OptimizeTable;
@@ -58,6 +59,7 @@ $application->add(new MigrateMailLogTokens());
 $application->add(new MigrateIpCreatedDayIndex());
 $application->add(new MigrateDropMailLogColumns());
 $application->add(new MigrateDropMailLogIndexes());
+$application->add(new MigrateMapUserConstraints());
 $application->add(new MigrateDb());
 
 $application->run();